requirements.txt
transformers 
torch 
Pillow
torchvision 
einops
accelerate
